+++ src/app/user/(center)/layout.tsx
@@ -1,33 +1,26 @@
-"use client";
-
 import PageTitle from "@/components/common/page-title";
+import SquareSideTabs from "@/components/common/square-side-tabs";
 import NavigationTabs from "@/components/user/navigation-tabs";
 import { userCenterRoutes } from "@/constants/routes";
-import { Metadata } from "next";
+import { type Metadata } from "next";
 import { usePathname } from "next/navigation";
 import React, { useMemo } from "react";
+
+export const metadata: Metadata = {
+  title: "我的空间 - CODEMATE",
+};
 
 const UserCenterLayout = ({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) => {
-  const pathname = usePathname();
-
-  const navItems = useMemo<NavItemType[]>(() => {
-    return userCenterRoutes?.map((r) => ({ ...r, isActive: pathname.startsWith(r.href) }));
-  }, [pathname]);
   return (
     <div className="m-auto max-w-screen-lg pb-10">
       <PageTitle>我的空间</PageTitle>
       <div className="flex gap-x-8">
-        <NavigationTabs items={navItems} />
-        <div className="flex-1">
-          <h3 className="text-[#3D3D3D] text-lg font-bold mb-5">
-            {navItems.find((r) => r.isActive)?.name ?? "我的中心"}
-          </h3>
-          {children}
-        </div>
+        <NavigationTabs />
+        <div className="flex-1">{children}</div>
       </div>
     </div>
   );
