+++ src/app/user/(center)/record/page.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import FaultList from "@/components/user/record/fault-list";
+import StarredList from "@/components/user/record/starred-list";
 import SubmitRecords from "@/components/user/record/submit-records";
 import { useUrlParam } from "@/hooks/useUrl";
 import { Tabs, type TabsProps } from "antd";
@@ -25,7 +26,7 @@ const tabItems: TabsProps["items"] = [
   {
     key: "favorites",
     label: "收藏",
-    children: null,
+    children: <StarredList />,
   },
 ];
